# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Citadel Agent API
  description: API for Citadel Agent - Autonomous Secure Workflow Engine
  version: 1.0.0
  contact:
    name: Citadel Agent Support
    email: support@citadel-agent.com
servers:
  - url: http://localhost:5001/api/v1
    description: Development server
  - url: https://api.citadel-agent.com/v1
    description: Production server
paths:
  /auth/login:
    post:
      summary: User login
      description: Login with email and password to get JWT token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: user@example.com
                password:
                  type: string
                  example: securepassword
              required:
                - email
                - password
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: user-123
        email:
          type: string
          format: email
          example: user@example.com
        username:
          type: string
          example: johndoe
        first_name:
          type: string
          example: John
        last_name:
          type: string
          example: Doe
        created_at:
          type: integer
          format: int64
          example: 1672531200
        updated_at:
          type: integer
          format: int64
          example: 1672531200
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Error message
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
...